의존성 추가 및 보드 로그인 추가

+++ src/main/java/com/hanghae0705/sbmoney/service/BoardService.java
@@ -1,24 +1,40 @@
 package com.hanghae0705.sbmoney.service;
 
+import com.auth0.jwt.interfaces.DecodedJWT;
 import com.hanghae0705.sbmoney.data.Message;
 import com.hanghae0705.sbmoney.model.domain.Board;
 import com.hanghae0705.sbmoney.model.domain.GoalItem;
 import com.hanghae0705.sbmoney.model.domain.User;
 import com.hanghae0705.sbmoney.repository.BoardRepository;
 import com.hanghae0705.sbmoney.repository.GoalItemRepository;
+import com.hanghae0705.sbmoney.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
+import static com.auth0.jwt.JWT.decode;
 
 @Service
 @RequiredArgsConstructor
 public class BoardService {
     private final BoardRepository boardRepository;
     private final GoalItemRepository goalItemRepository;
+    private final UserRepository userRepository;
+    private Optional<User> getUser(String authorization){
+        String token = authorization.substring(7);
+        DecodedJWT decodeToken = decode(token);
+        String username = decodeToken.getClaim("sub").toString();
+        int length = username.length();
+        username = username.substring(1,length-1);
+        Optional<User> user = userRepository.findByUsername(username);
 
+        return user;
+
+    }
     @Transactional
     public Message GetBoard() {
         List<Board> boardList = boardRepository.findAll();
@@ -31,8 +47,8 @@ public class BoardService {
     }
 
     @Transactional
-    public Message postBoard(Board.Request request) {
-        User user = null;
+    public Message postBoard(Board.Request request,String authorization) {
+        Optional<User> user = getUser(authorization);
         GoalItem goalItem = goalItemRepository.findAllById(request.getGoalItemId());
         Board board = new Board(request, goalItem, user);
         boardRepository.save(board);
@@ -40,18 +56,30 @@ public class BoardService {
     }
 
     @Transactional
-    public Message putBoard(Board.Update request, Long boardId) {
+    public Message putBoard(Board.Update request, Long boardId,String authorization) {
+        Optional<User> user = getUser(authorization);
         Board board = boardRepository.findAllById(boardId);
-        board.updateBoard(request);
-        return new Message(true, "게시글을 수정하였습니다");
+        if(user.get().getId().equals(board.getUser().getId())){
+            board.updateBoard(request);
+            return new Message(true, "게시글을 수정하였습니다");
+        }else {
+            return new Message(false, "게시글을 수정에 실패하였습니다");
+        }
+
+
 
     }
 
     @Transactional
-    public Message deleteBoard(Long boardId) {
+    public Message deleteBoard(Long boardId,String authorization) {
+        Optional<User> user = getUser(authorization);
         Board board = boardRepository.findAllById(boardId);
-        boardRepository.delete(board);
-        return new Message(true, "게시글을 삭제하였습니다");
+        if(user.get().getId().equals(board.getUser().getId())){
+            boardRepository.delete(board);
+            return new Message(true, "게시글을 삭제하였습니다");
+        }else {
+            return new Message(false, "게시글을 삭제에 실패하였습니다");
+        }
     }
